requirements.txt
tqdm
bayoo-docx
Flask
tenacity
openai